requirements.txt
asgiref==3.8.1
astroid==3.3.9
colorama==0.4.6
defusedxml==0.7.1
dill==0.4.0
Django==4.2.19
djangorestframework==3.16.0
djangorestframework-xml==2.0.0
greenlet==3.1.1
isort==6.0.1
mccabe==0.7.0
mssql==1.0.1
mssql-django==1.5
mysqlclient==2.2.7
pillow==11.2.1
platformdirs==4.3.7
pylint==3.3.6
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
pyodbc==5.2.0
pytz==2025.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.1
